Migrate raycast first function
